backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
pydantic==2.8.0

# ── Shoonya API Integration (optional) ──
NorenRestApiPy>=0.0.23
pyotp>=2.9.0
python-dotenv>=1.0.0
